Address review comments: rename to snake_case, add hypothesis pin, update docs
- Rename larkInmantaParser.py -> lark_parser.py and larkInmanta.lark -> inmanta.lark
  to follow Python naming conventions (sanderr)
- Add hypothesis==6.151.9 pin to requirements.txt (sanderr)
- Clarify map_lookup comment in README (sanderr)
- Remove fast-exit rows from README performance table (moved to PR #10091)
- Update vargs/dispatch comment to explain rationale
- Update all import references and MANIFEST.in
