+++ MANIFEST.in
@@ -3,7 +3,7 @@ include tox.ini
 include setup.py
 include setup.cfg
 include src/inmanta/py.typed
-include src/inmanta/parser/larkInmanta.lark
+include src/inmanta/parser/inmanta.lark
 include src/inmanta/protocol/auth/default_policy.rego
 
 # explainers for compiler exceptions
