+++ requirements.txt
@@ -14,6 +14,7 @@ docstring-parser==0.17.0
 email-validator==2.3.0
 furo==2025.12.19
 graphviz==0.21
+hypothesis==6.151.9
 inmanta-dev-dependencies==2.192.0
 inmanta-module-std
 inmanta-sphinx==2.3.0
